+++ docs/changelog.rst
@@ -17,6 +17,14 @@ New features
 ------------
 
 * Importer for Berlin fsc2 format: :class:`trepr.io.Fsc2Importer`
+* :class:`trepr.analysis.MWFrequencyValues` to extract microwave frequency values from a dataset as a new, calculated dataset (e.g., for graphical display)
+* :class:`trepr.analysis.MWFrequencyDrift` can return (calculated) datasets
+
+
+Changes
+-------
+
+* :class:`trepr.analysis.MWFrequencyDrift` renamed from ``MwFreqAnalysis``
 
 
 Fixes
@@ -25,6 +33,7 @@ Fixes
 * :class:`trepr.processing.BackgroundCorrection` subtracts background
 * :class:`trepr.io.TezImporter` gets mapper from package data
 * :class:`trepr.io.DatasetImporterFactory` falls back to ASpecD-supported formats if no matching format is found.
+* :class:`trepr.analysis.MWFrequencyDrift` returns correct value for ratio of drift to magnetic field axis step width
 
 
 Version 0.1.0